Replace the Bulk node with a more succinct one
This is from my original `pyiron_atomistics` node package. The change is functionally equivalent, except that the input for the new version will now dynamically evolve if it changes in the upstream dependency (`pyiron_atomistics`), the import might possibly slow us down, and it's a lot less lines. My main motivation for this change is to keep an example of directly wrapping an existing function without re-defining the IO.